restructuring number system performance measurement approach to add reuse
