+++ tests/performance/16b_posit.cpp
@@ -10,7 +10,7 @@
 // second: disable posit arithmetic exceptions
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 0
 #include <universal/posit/posit>
-#include "posit_performance.hpp"
+#include <universal/performance/number_system.hpp>
 
 int main(int argc, char** argv)
 try {
